+++ extensions/Worker.Extensions.CosmosDB/release_notes.md
@@ -1,5 +1,9 @@
-## Release notes
+## What's Changed
+
 <!-- Please add your release notes in the following format:
 - My change description (#PR/#issue)
 -->
-- Fixed incorrect type of CosmosDBTriggerAttribute's `StartFromTime` property.
+
+### Microsoft.Azure.Functions.Worker.Extensions.CosmosDB <version>
+
+- Add `BindingCapabilities` attribute to CosmosDb trigger to express function-level retry capabilities. (#1457)