+++ extensions/Worker.Extensions.Timer/release_notes.md
@@ -1,4 +1,9 @@
-## Release notes
+## What's Changed
+
 <!-- Please add your release notes in the following format:
 - My change description (#PR/#issue)
 -->
+
+### Microsoft.Azure.Functions.Worker.Extensions.Timer <version>
+
+- Add `BindingCapabilities` attribute to  TimerTrigger to express function-level retry capabilities. (#1457)